refactor error organization
