+++ gene/etl/__init__.py
@@ -1,4 +1,20 @@
 """Module to load and init namespace at package level."""
-from .ensembl import Ensembl  # noqa: F401
-from .hgnc import HGNC  # noqa: F401
-from .ncbi import NCBI  # noqa: F401
+from .ensembl import Ensembl
+from .exceptions import (
+    FileVersionError,
+    NormalizerEtlError,
+    SourceFetchError,
+    SourceFormatError,
+)
+from .hgnc import HGNC
+from .ncbi import NCBI
+
+__all__ = [
+    "Ensembl",
+    "HGNC",
+    "NCBI",
+    "NormalizerEtlError",
+    "FileVersionError",
+    "SourceFetchError",
+    "SourceFormatError",
+]
