+++ gene/etl/base.py
@@ -192,23 +192,3 @@ class Base(ABC):
         if not Path(seqrepo_dir).exists():
             raise NotADirectoryError(f"Could not find {seqrepo_dir}")
         return SeqRepo(seqrepo_dir)
-
-
-class NormalizerEtlError(Exception):
-    """Base ETL exception."""
-
-
-class FileVersionError(NormalizerEtlError):
-    """Raise when unable to parse version number from saved data file."""
-
-
-class SourceFormatError(NormalizerEtlError):
-    """Raise when source data formatting is incompatible with the source transformation
-    methods: for example, if columns in a CSV file have changed.
-    """
-
-
-class SourceFetchError(NormalizerEtlError):
-    """Raise during data acquisition when data fetch fails (e.g. unable to get latest
-    version number, or connection failure during download)
-    """
