+++ gene/etl/ensembl.py
@@ -13,8 +13,8 @@ from gffutils.feature import Feature
 
 from gene import APP_ROOT
 from gene.database import AbstractDatabase
-from gene.etl.base import (
-    Base,
+from gene.etl.base import Base
+from gene.etl.exceptions import (
     FileVersionError,
     NormalizerEtlError,
     SourceFetchError,
